Split Field per component, and correct a wrong claim about the bundle
Field.tsx now has one module per control -- Input, Textarea, NumberInput,
SearchInput, Select -- with the shared accessibility scaffolding in
field-internals.tsx so they still cannot drift apart on label association,
description linking or invalid state. Field.tsx remains as a compatibility
re-export.

The correction
--------------
The previous commit and docs/HANDOFF.md predicted, as fact, that this split
would be "the next reduction" and would bring /login toward its 30kB target.
That was wrong. Measured before and after:

  before: /login 50.8kB · /register 43.7kB
  after:  /login 50.8kB · /register 43.7kB

No measurable change. The same reasoning that correctly ruled out Feedback
(measured, 0.1kB) was applied to Field WITHOUT measuring first, and it did not
hold. The split is kept -- one module per control is better hygiene and costs
nothing -- but it is not credited with a saving it did not make.

What the bundle actually contains
---------------------------------
Read from .next/app-build-manifest.json:

  11.4kB  design system, SHARED by /login and /register -- paid once
  10.0kB  /login page code
   3.0kB  /register page code

So the design system's marginal cost is 11.4kB shared, not the ~23kB the
per-route budget delta implies. The budget metric sums every chunk a route
loads, attributing shared code to each route that touches it, which makes the
first migrations look worse than the steady state because there is nothing yet
to amortise against.

That is a prediction, and this time it is recorded as one: each additional
migrated route should add its own page code and little else. If per-route JS
keeps climbing ~20kB per route as the third and fourth land, this explanation
is wrong too and the real cause is still unfound.

The auth pages now import from @/components/ds/Input directly.

Verified: lint, typecheck, 76 test files / 1892 tests, coverage thresholds met,
build 76/76 pages, bundle budgets.
