+++ src/app/register/page.tsx
@@ -7,7 +7,7 @@ import { useState } from "react";
 // re-exports Modal, which pulls focus-trap-react into any page that touches
 // the barrel. That alone took /login from 28kB to 121kB of route JS.
 import { Button } from "@/components/ds/Button";
-import { Input } from "@/components/ds/Field";
+import { Input } from "@/components/ds/Input";
 import { Banner } from "@/components/ds/Banner";
 import { AuthShell, AuthStatus, AuthActions, codeInputClass } from "@/components/ds/AuthShell";
